\name{hclust}
\alias{hclust-class}
\docType{class}
\title{Class "hclust"}
\description{
  Creates an S4 class equivalent to the S3 hclust class returned by the
  \code{hclust} function.
}
\author{
  Kevin R. Coombes \email{krc@silicovore.com}
}
\seealso{
  \code{\link[stats]{cutree}},
  \code{\link[stats]{hclust}},
  \code{\link[stats]{heatmap}}
}
\keyword{classes}

