\name{hclust}
\alias{hclust-class}
\docType{class}
\title{Class "hclust"}
\description{
  Creates an S4 class equivalent to the S3 hclust class returned by the
  \code{hclust} function.
}
\author{
  Kevin R. Coombes \email{kcoombes@mdanderson.org}
}
\seealso{
  \code{\link{cutree}},
  \code{\link{hclust}},
  \code{\link{heatmap}}
}
\keyword{classes}

