\name{hclust}
\docType{class}
\alias{hclust-class}

\title{An S4 hclust class}
\description{ Creates an S4 hclust class that is equivalent to the S3
  hclust class returned by the hclust function.}
\author{Kevin R. Coombes <kcoombes@mdanderson.org>}
\seealso{ \code{\link{heatmap}}, \code{\link{hclust}}, \code{\link{cutree}} }
\keyword{classes}
